SQLAgent/requirements.txt
ollama
sqlalchemy
psycopg2-binary
llama-index-core
llama-index-embeddings-huggingface
transformers
torch
numpy